requirements.txt
tensorflow>=2.6
pandas>=1.1
numpy>=1.19
matplotlib>=3.4
sklearn
pyyaml